Fix Makefile, e2e tests, naming
This commit fixes minor errors in the makefile, due to the absence of
CRDs in the project.

It also includes some changes on the e2e test suite, that make them more
stable and repeatable.

The naming of the k8s resources was slightly improved too, but more
polishing might be needed

Assisted by: Cursor (powered by Claude 4 Sonnet)
Ref: https://issues.redhat.com/browse/EC-1317