Implement pipelinerun-controller business logic
Whenever a PipelineRun is completed and signed, the controller will
detect it, and will trigger a Conforma TaskRun.
It will also add to the PipelineRun the annotation
'conforma.dev/triggered-on', to mark it as already processed.

Added unit and e2e test coverage.

Assisted by: Cursor (powered by Claude 4 Sonnet)

Ref: https://issues.redhat.com/browse/EC-1317
Ref: https://issues.redhat.com/browse/EC-1319